pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "graphnlp-intel"
version = "0.1.0"
description = "Hybrid Graph-NLP Intelligence Platform"
requires-python = ">=3.10"
dependencies = [
  "spacy>=3.7",
  "transformers>=4.40",
  "sentence-transformers>=3.0",
  "networkx>=3.3",
  "torch>=2.3",
  "torch-geometric>=2.5",
  "pyvis>=0.3",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
  "pydantic-settings>=2.2",
  "neo4j>=5.20",
  "redis>=5.0",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "pdfminer.six>=20231228",
  "pyyaml>=6.0",
  "typer>=0.12",
  "python-louvain>=0.16",
  "vaderSentiment>=3.3",
  "numpy>=1.26",
  "scikit-learn>=1.4",
]

[project.optional-dependencies]
serve  = ["celery[redis]>=5.4", "flower>=2.0", "aiokafka>=0.11"]
sonar  = ["sonar-space>=0.2"]
dev    = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "pre-commit>=3",
  "fakeredis>=2.23",
  "ruff>=0.4",
  "mypy>=1.10",
]

[project.scripts]
graphnlp = "graphnlp.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
